refactored geolocation check, added ntp and dns checks, added variables to not start processes

+++ processor.py
@@ -4,7 +4,7 @@ from detections import update_LOCAL_NETWORKS, detect_geolocation_flows, detect_n
 from notifications import send_test_telegram_message  # Import send_test_telegram_message from notifications.py
 from integrations.maxmind import create_geolocation_db, load_geolocation_data
 from utils import log_info, log_warn, log_error  # Import log_info from utils
-from const import CONST_PROCESSING_INTERVAL, IS_CONTAINER, CONST_NEWFLOWS_DB, CONST_ALLFLOWS_DB, CONST_ALERTS_DB, CONST_WHITELIST_DB, CONST_CONFIG_DB, CONST_CREATE_WHITELIST_SQL, CONST_CREATE_ALERTS_SQL, CONST_CREATE_ALLFLOWS_SQL, CONST_CREATE_NEWFLOWS_SQL, CONST_CREATE_CONFIG_SQL
+from const import CONST_SCHEDULE_PROCESSOR, CONST_CLEAN_NEWFLOWS,CONST_REINITIALIZE_DB,CONST_PROCESSING_INTERVAL, IS_CONTAINER, CONST_NEWFLOWS_DB, CONST_ALLFLOWS_DB, CONST_ALERTS_DB, CONST_WHITELIST_DB, CONST_CONFIG_DB, CONST_CREATE_WHITELIST_SQL, CONST_CREATE_ALERTS_SQL, CONST_CREATE_ALLFLOWS_SQL, CONST_CREATE_NEWFLOWS_SQL, CONST_CREATE_CONFIG_SQL
 import schedule
 import time
 import logging
@@ -15,6 +15,9 @@ import os
 
 if (IS_CONTAINER):
     PROCESSING_INTERVAL=os.getenv("PROCESSING_INTERVAL", CONST_PROCESSING_INTERVAL)
+    REINITIALIZE_DB=os.getenv("REINITIALIZE_DB", CONST_REINITIALIZE_DB)
+    CLEAN_NEWFLOWS=os.getenv("CLEAN_NEWFLOWS", CONST_CLEAN_NEWFLOWS)
+    SCHEDULE_PROCESSOR=os.getenv("SCHEDULE_PROCESSOR", CONST_SCHEDULE_PROCESSOR)
 
 # Function to process data
 def process_data():
@@ -33,7 +36,8 @@ def process_data():
             rows = cursor.fetchall()
 
             log_info(logger, f"[INFO] Fetched {len(rows)} rows from the database.")
-            delete_all_records(CONST_NEWFLOWS_DB)
+            if (CLEAN_NEWFLOWS):
+                delete_all_records(CONST_NEWFLOWS_DB)
 
             # Pass the rows to update_all_flows
             update_allflows(rows, config_dict)
@@ -79,10 +83,14 @@ if __name__ == "__main__":
     geolocation_data = load_geolocation_data()
 
     send_test_telegram_message()
-    #delete databases to start clean - temporary
-    #delete_database(CONST_ALERTS_DB)
     delete_database(CONST_CONFIG_DB)
-    #delete_database(CONST_ALLFLOWS_DB)
+
+    if (REINITIALIZE_DB):
+        delete_database(CONST_NEWFLOWS_DB)
+        delete_database(CONST_ALLFLOWS_DB)
+        delete_database(CONST_ALERTS_DB)
+        delete_database(CONST_WHITELIST_DB)
+        delete_database(CONST_CONFIG_DB)
 
     # Initialize all required databases
     create_database(CONST_NEWFLOWS_DB, CONST_CREATE_NEWFLOWS_SQL)
@@ -95,6 +103,7 @@ if __name__ == "__main__":
     init_configurations()
     process_data()
 
-    while True:
-        schedule.run_pending()
-        time.sleep(1)
+    if (SCHEDULE_PROCESSOR):
+        while True:
+            schedule.run_pending()
+            time.sleep(1)
